#30 refactor: 모바일 화면에서도 같은 로직을 활용하기 위해 useReservation 커스텀 훅 제작
